a substantial API cleanup and refactor.
New class "Client" represents a client connection, which sends requets
and can be sent responses.
New class "Handler" is a base class for handlers, of which WSGIHandler
is now a subclass.
Several methods shuffled, renamed and fudged to make use of these new
classes; see the ChangeLog for the details.